+++ dev/NetMath.js
@@ -120,6 +120,44 @@ class NetMath {
         }
     }
 
+    // Weights init
+    static uniform (size, {limit}) {
+        return [...new Array(size)].map(v => Math.random()*2*limit-limit)
+    }
+
+    static gaussian (size, {mean, stdDeviation}) {
+        return [...new Array(size)].map(() => {
+            // Polar Box Muller
+            let x1, x2, r, y
+
+            do {
+                x1 = 2 * Math.random() -1
+                x2 = 2 * Math.random() -1
+                r = x1**2 + x2**2
+            } while (r >= 1 || !r)
+
+            return mean + (x1 * (Math.sqrt(-2 * Math.log(r) / r))) * stdDeviation
+        })
+    }
+
+    static xavierNormal (size, {fanIn, fanOut}) {
+        return fanOut || fanOut==0 ? NetMath.gaussian(size, {mean: 0, stdDeviation: Math.sqrt(2/(fanIn+fanOut))})
+                                   : NetMath.lecunNormal(size, {fanIn})
+    }
+
+    static xavierUniform (size, {fanIn, fanOut}) {
+        return fanOut || fanOut==0 ? NetMath.uniform(size, {limit: Math.sqrt(6/(fanIn+fanOut))})
+                                   : NetMath.lecunUniform(size, {fanIn})
+    }    
+
+    static lecunNormal (size, {fanIn}) {
+        return NetMath.gaussian(size, {mean: 0, stdDeviation: Math.sqrt(1/fanIn)})
+    }
+
+    static lecunUniform (size, {fanIn}) {
+        return NetMath.uniform(size, {limit: Math.sqrt(3/fanIn)})
+    }
+
     // Other
     static softmax (values) {
         const total = values.reduce((prev, curr) => prev+curr, 0)
@@ -128,6 +166,12 @@ class NetMath {
 
     static sech (value) {
         return (2*Math.exp(-value))/(1+Math.exp(-2*value))
+    }
+
+    static standardDeviation (arr) {
+        const avg = arr.reduce((p,c) => p+c) / arr.length
+        const diffs = arr.map(v => v - avg).map(v => v**2)
+        return Math.sqrt(diffs.reduce((p,c) => p+c) / diffs.length)
     }
 
     static maxNorm () {